+++ src/click/types.py
@@ -107,7 +107,7 @@ class ParamType:
         """Helper method to fail with an invalid value message."""
         raise BadParameter(message, ctx=ctx, param=param)
 
-    def shell_complete(self, ctx, args, incomplete):
+    def shell_complete(self, ctx, param, args, incomplete):
         """Return a list of
         :class:`~click.shell_completion.CompletionItem` objects for the
         incomplete value. Most types do not provide completions, but
@@ -115,6 +115,7 @@ class ParamType:
         completions as well.
 
         :param ctx: Invocation context for this command.
+        :param param: The parameter that is requesting completion.
         :param args: List of complete args before the incomplete value.
         :param incomplete: Value being completed. May be empty.
 
@@ -264,10 +265,11 @@ class Choice(ParamType):
     def __repr__(self):
         return f"Choice({list(self.choices)})"
 
-    def shell_complete(self, ctx, args, incomplete):
+    def shell_complete(self, ctx, param, args, incomplete):
         """Complete choices that start with the incomplete value.
 
         :param ctx: Invocation context for this command.
+        :param param: The parameter that is requesting completion.
         :param args: List of complete args before the incomplete value.
         :param incomplete: Value being completed. May be empty.
 
@@ -612,11 +614,12 @@ class File(ParamType):
                 ctx,
             )
 
-    def shell_complete(self, ctx, args, incomplete):
+    def shell_complete(self, ctx, param, args, incomplete):
         """Return a special completion marker that tells the completion
         system to use the shell to provide file path completions.
 
         :param ctx: Invocation context for this command.
+        :param param: The parameter that is requesting completion.
         :param args: List of complete args before the incomplete value.
         :param incomplete: Value being completed. May be empty.
 
@@ -757,12 +760,13 @@ class Path(ParamType):
 
         return self.coerce_path_result(rv)
 
-    def shell_complete(self, ctx, args, incomplete):
+    def shell_complete(self, ctx, param, args, incomplete):
         """Return a special completion marker that tells the completion
         system to use the shell to provide path completions for only
         directories or any paths.
 
         :param ctx: Invocation context for this command.
+        :param param: The parameter that is requesting completion.
         :param args: List of complete args before the incomplete value.
         :param incomplete: Value being completed. May be empty.